[Issue 204] Explicitly use PKCS12 store type for TCK tests
The SSL TCK tests use PKCS12 key/trust store types, but the
tests themselves either specify JKS or rely on the default
store type (for JDK 8, that is JKS - for 9+, that is PKCS12).
Most JDKs can handle PKCS12 types, even when the code
specifies JKS, but IBM JDK 8 cannot.

This change explicitly sets the store type in JVM system
properties, the Jetty test code and MP Config properties
for the test application.

Signed-off-by: Andy McCright <j.andrew.mccright@gmail.com>